+++ resources/lang/en/api-service.php
@@ -2,5 +2,5 @@
 
 // translations for Rupadana/ApiService
 return [
-    //
+
 ];