+++ src/ApiService.php
@@ -2,16 +2,19 @@
 
 namespace Rupadana\ApiService;
 
+use Filament\Panel;
 use Illuminate\Routing\Router;
 use Illuminate\Support\Facades\Route;
+use Rupadana\ApiService\Concerns\HasTenancy;
 
 class ApiService
 {
+    use HasTenancy;
+
     /**
      * Filament Resource
      */
     protected static ?string $resource = null;
-
     protected static ?string $groupRouteName = null;
 
     /**
@@ -30,7 +33,7 @@ class ApiService
         return static::$resource;
     }
 
-    public static function registerRoutes()
+    public static function registerRoutes(Panel $panel)
     {
 
         $slug = static::getResource()::getSlug();
@@ -39,9 +42,10 @@ class ApiService
             ->replace('/', '.')
             ->append('.');
 
-        Route::name(
-            $name
-        )
+        $resourceRouteMiddlewares = static::useResourceMiddlewares() ? static::getResource()::getRouteMiddleware($panel) : [];
+
+        Route::name($name)
+            ->middleware($resourceRouteMiddlewares)
             ->prefix(static::$groupRouteName ?? $slug)
             ->group(function (Router $route) {
                 static::handlers();
@@ -55,5 +59,15 @@ class ApiService
     public static function handlers(): array
     {
         return [];
+    }
+
+    public static function isRoutePrefixedByPanel(): bool
+    {
+        return config('api-service.route.panel_prefix', true);
+    }
+
+    public static function useResourceMiddlewares(): bool
+    {
+        return config('api-service.route.use_resource_middlewares', false);
     }
 }